+++ nata-monkey/src/main/java/com/nata/utils/HttpUtil.java
@@ -7,11 +7,14 @@ import org.apache.http.client.methods.HttpPost;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.message.BasicNameValuePair;
+import org.apache.http.protocol.HTTP;
 
 import java.io.IOException;
 import java.lang.reflect.Array;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Author: Calvin Meng
@@ -20,43 +23,51 @@ import java.util.List;
  */
 public class HttpUtil {
 
-    public static void post(String url,ArrayList<String> messages) throws IOException {
+    public static void post(String url,HashMap<String,String> props) throws IOException {
         CloseableHttpClient client = HttpClients.createDefault();
         HttpPost httpPost = new HttpPost(url);
+//        httpPost.addHeader("charset", HTTP.UTF_8);
 
         List<NameValuePair> params = new ArrayList<>();
-        for (String message: messages
-             ) {
-            params.add(new BasicNameValuePair("message", message));
+        for (Map.Entry<String,String> entry: props.entrySet()) {
+            params.add(new BasicNameValuePair(entry.getKey(), entry.getValue()));
         }
 
-        httpPost.setEntity(new UrlEncodedFormEntity(params));
+        httpPost.setEntity(new UrlEncodedFormEntity(params,"UTF-8"));
 
         CloseableHttpResponse response = client.execute(httpPost);
         client.close();
     }
 
-
-    public static void postTestResult(ArrayList<String> messages) throws IOException {
-        post("http://localhost:3000/test",messages);
+    public static void postSummary(HashMap<String,String> props) throws IOException {
+        String summaryUrl = "http://localhost:3000/summary";
+        post(summaryUrl,props);
     }
 
-    public static void postLive(String message) throws IOException {
-        ArrayList<String> messages = new ArrayList<>();
-        messages.add(message);
-        post("http://localhost:3000/live",messages);
+    public static void postAction(String message) throws IOException {
+        HashMap<String,String > messages = new HashMap<>();
+        messages.put("message",message);
+        post("http://localhost:3000/action",messages);
     }
 
-    public static void postSummary(String summary) throws IOException {
-        ArrayList<String> messages = new ArrayList<>();
-        messages.add(summary);
-        post("http://localhost:3000/summary",messages);
+    public static void postState(String message) throws IOException {
+        HashMap<String,String > messages = new HashMap<>();
+        messages.put("message",message);
+        post("http://localhost:3000/state",messages);
     }
 
-    public static void postActivity(String activity) throws IOException {
-        ArrayList<String> messages = new ArrayList<>();
-        messages.add(activity);
+    public static void postActivity(String message) throws IOException {
+        HashMap<String,String > messages = new HashMap<>();
+        messages.put("message",message);
         post("http://localhost:3000/activity",messages);
     }
+
+    public static void postWidget(String message) throws IOException {
+        HashMap<String,String > messages = new HashMap<>();
+        messages.put("message",message);
+        post("http://localhost:3000/widget",messages);
+    }
+
+
 
 }